// yugabyted-ui/ui/src/api/src/models/SlowQueryResponseYSQLQueryItem.ts
// tslint:disable
/**
 * Yugabyte Cloud
 * YugabyteDB as a Service
 *
 * The version of the OpenAPI document: v1
 * Contact: support@yugabyte.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */




/**
 * Schema for Slow Query Response YSQL Query Item
 * @export
 * @interface SlowQueryResponseYSQLQueryItem
 */
export interface SlowQueryResponseYSQLQueryItem  {
  /**
   * 
   * @type {number}
   * @memberof SlowQueryResponseYSQLQueryItem
   */
  queryid?: number;
  /**
   * 
   * @type {string}
   * @memberof SlowQueryResponseYSQLQueryItem
   */
  query?: string;
  /**
   * 
   * @type {string}
   * @memberof SlowQueryResponseYSQLQueryItem
   */
  rolname?: string;
  /**
   * 
   * @type {string}
   * @memberof SlowQueryResponseYSQLQueryItem
   */
  datname?: string;
  /**
   * 
   * @type {number}
   * @memberof SlowQueryResponseYSQLQueryItem
   */
  calls?: number;
  /**
   * 
   * @type {number}
   * @memberof SlowQueryResponseYSQLQueryItem
   */
  local_blks_hit?: number;
  /**
   * 
   * @type {number}
   * @memberof SlowQueryResponseYSQLQueryItem
   */
  local_blks_written?: number;
  /**
   * 
   * @type {number}
   * @memberof SlowQueryResponseYSQLQueryItem
   */
  max_time?: number;
  /**
   * 
   * @type {number}
   * @memberof SlowQueryResponseYSQLQueryItem
   */
  mean_time?: number;
  /**
   * 
   * @type {number}
   * @memberof SlowQueryResponseYSQLQueryItem
   */
  min_time?: number;
  /**
   * 
   * @type {number}
   * @memberof SlowQueryResponseYSQLQueryItem
   */
  rows?: number;
  /**
   * 
   * @type {number}
   * @memberof SlowQueryResponseYSQLQueryItem
   */
  stddev_time?: number;
  /**
   * 
   * @type {number}
   * @memberof SlowQueryResponseYSQLQueryItem
   */
  total_time?: number;
}